removing ubuntu from ci matrix for now

+++ tools/install_dependencies.sh
@@ -123,18 +123,20 @@ elif [ -x "$(command -v apt-get)" ]; then
     # Install a version of CMake more up-to-date than what is available by
     # default for Ubuntu Bionic
 
-    sudo apt purge --auto-remove cmake
-    version=3.16
-    build=2
-    mkdir -p ~/temp
-    pushd ~/temp
-      wget https://cmake.org/files/v$version/cmake-$version.$build.tar.gz
-      tar -xzvf cmake-$version.$build.tar.gz
-      cd cmake-$version.$build/
-      ./bootstrap
-      make -j
-      sudo make install
-    popd
+    if [[ ! -z $TRAVIS ]]; then
+      sudo apt purge --auto-remove cmake
+      version=3.16
+      build=2
+      mkdir -p ~/temp
+        pushd ~/temp
+          wget https://cmake.org/files/v$version/cmake-$version.$build.tar.gz
+          tar -xzvf cmake-$version.$build.tar.gz
+          cd cmake-$version.$build/
+          ./bootstrap
+          make -j
+          sudo make install
+        popd
+    fi
 
     sudo apt-get install \
         gcc-9 \
